refactor: available dog 리팩토링
